refactor and test fixes
